Refactor router to use sync callbacks for easier start of the project
