+++ src/controllers/UserController.ts
@@ -9,7 +9,7 @@ export class UserController {
 	constructor(private userService: UserService) {}
 
 	// POST /api/v1/auth/sign-in"
-	public async signUser(req: Request, res: Response): Promise<Response> {
-		return await this.userService.tryToSign(req.body, req.headers, res);
+	public signUser(req: Request, res: Response): Response {
+		return this.userService.tryToSign(req.body, req.headers, res);
 	}
 }
