+++ src/dtos/HttpErrorResponseDTO.ts
