+++ src/routes/auth.route.ts
@@ -6,8 +6,8 @@ import { UserController } from 'src/controllers/UserController';
 const router = Router();
 const userController: UserController = injector.get(UserController);
 
-router.post('/sign-in', async (req: Request, res: Response) => {
-		return await userController.signUser(req, res);
+router.post('/sign-in', (req: Request, res: Response) => {
+		return userController.signUser(req, res);
 	}
 );
